Instances of client_t in the server, for the core engine, will now have a reference to the network protocol version in use by the respective clients, which could be different from the version in the server. Fix for bug in server, which upon receiving an acknowledgement of a protocol upgrade from one of the clients, assumed the upgrade applied to *all clients*, thus allowing the server to send incorrect commands to clients that did not support them.

+++ id1/pr_edict.cpp
@@ -1187,12 +1187,12 @@ edict_t *EDICT_NUM(int n)
 
 int NUM_FOR_EDICT(edict_t *e)
 {
-	int		b;
+	intptr_t		b;
 	
 	b = (byte *)e - (byte *)sv.edicts.data();
 	b = b / pr_edict_size;
 	
 	if (b < 0 || b >= sv.num_edicts)
 		Sys_Error ("NUM_FOR_EDICT: bad pointer");
-	return b;
+	return (int)b;
 }
